Changed telemetry parameters from Int to Double

+++ src/main/java/com/github/jbladejs/iot/Driver.kt
@@ -1,8 +1,8 @@
 package com.github.jbladejs.iot
 
 interface Driver {
-    fun getLightIntensity() : Int
-    fun getEnergyUsage() : Int
+    fun getLightIntensity() : Double
+    fun getEnergyUsage() : Double
     fun isLightOn() : Boolean
     fun turnOnTheLight()
     fun turnOffTheLight()
