+++ src/main/java/com/github/jbladejs/iot/tools/TelemetryData.kt
@@ -2,6 +2,6 @@ package com.github.jbladejs.iot.tools
 
 import com.google.gson.Gson
 
-internal data class TelemetryData(val energyUsage: Int, val lightIntensity: Int) {
+internal data class TelemetryData(val energyUsage: Double, val lightIntensity: Double) {
     fun serialize() : String = Gson().toJson(this)
 }
